// book-network-ui/src/app/services/model/borrowedBookResponse.ts
/**
 * OpenApi specification - Jayarrakesh
 *
 * Contact: prabakars2003@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BorrowedBookResponse { 
    id?: number;
    title?: string;
    authorname?: string;
    isbn?: string;
    rate?: number;
    returned?: boolean;
    returnApproved?: boolean;
}