// book-network-ui/src/app/services/model/pageResponseBorrowedBookResponse.ts
/**
 * OpenApi specification - Jayarrakesh
 *
 * Contact: prabakars2003@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BorrowedBookResponse } from './borrowedBookResponse';


export interface PageResponseBorrowedBookResponse { 
    content?: Array<BorrowedBookResponse>;
    number?: number;
    size?: number;
    totalElements?: number;
    totalPages?: number;
    first?: boolean;
    last?: boolean;
}