// book-network-ui/src/app/services/model/feedbackRequest.ts
/**
 * OpenApi specification - Jayarrakesh
 *
 * Contact: prabakars2003@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FeedbackRequest { 
    note?: number;
    comment: string;
    bookId: number;
}